+++ sdk/core/azure-core/src/http/transport_policy.cpp
@@ -15,6 +15,7 @@ using Azure::Core::Context;
 using namespace Azure::Core::IO;
 using namespace Azure::Core::Http;
 using namespace Azure::Core::Http::Policies;
+using namespace Azure::Core::Http::Policies::_internal;
 
 std::shared_ptr<HttpTransport> Azure::Core::Http::Policies::_detail::GetTransportAdapter()
 {
@@ -57,7 +58,7 @@ std::unique_ptr<RawResponse> TransportPolicy::Send(
   // If ReadToEnd fail, retry policy will eventually call this again
   // Using DownloadViaStream and getting an error code would also get to here to download error from
   // body
-  auto bodyStream = response->GetBodyStream();
+  auto bodyStream = response->ExtractBodyStream();
   response->SetBody(bodyStream->ReadToEnd(ctx));
   // BodyStream is moved out of response. This makes transport implementation to clean any active
   // session with sockets or internal state.