+++ sdk/core/azure-core/test/ut/curl_options.cpp
@@ -36,7 +36,8 @@ namespace Azure { namespace Core { namespace Test {
     auto transportAdapter = std::make_shared<Azure::Core::Http::CurlTransport>(curlOptions);
     Azure::Core::Http::Policies::TransportOptions options;
     options.Transport = transportAdapter;
-    auto transportPolicy = std::make_unique<Azure::Core::Http::Policies::TransportPolicy>(options);
+    auto transportPolicy
+        = std::make_unique<Azure::Core::Http::Policies::_internal::TransportPolicy>(options);
 
     std::vector<std::unique_ptr<Azure::Core::Http::Policies::HttpPolicy>> policies;
     policies.emplace_back(std::move(transportPolicy));
@@ -66,7 +67,8 @@ namespace Azure { namespace Core { namespace Test {
     auto transportAdapter = std::make_shared<Azure::Core::Http::CurlTransport>(curlOptions);
     Azure::Core::Http::Policies::TransportOptions options;
     options.Transport = transportAdapter;
-    auto transportPolicy = std::make_unique<Azure::Core::Http::Policies::TransportPolicy>(options);
+    auto transportPolicy
+        = std::make_unique<Azure::Core::Http::Policies::_internal::TransportPolicy>(options);
 
     std::vector<std::unique_ptr<Azure::Core::Http::Policies::HttpPolicy>> policies;
     policies.emplace_back(std::move(transportPolicy));
@@ -100,7 +102,7 @@ namespace Azure { namespace Core { namespace Test {
 
     auto transportAdapter = std::make_shared<Azure::Core::Http::CurlTransport>(curlOptions);
     auto transportPolicy =
-  std::make_unique<Azure::Core::Http::Policies::TransportPolicy>(transportAdapter);
+  std::make_unique<Azure::Core::Http::Policies::_internal::TransportPolicy>(transportAdapter);
 
     std::vector<std::unique_ptr<Azure::Core::Http::Policies::HttpPolicy>> policies;
     policies.emplace_back(std::move(transportPolicy));
@@ -126,7 +128,7 @@ namespace Azure { namespace Core { namespace Test {
 
     auto transportAdapter = std::make_shared<Azure::Core::Http::CurlTransport>(curlOptions);
     auto transportPolicy =
-  std::make_unique<Azure::Core::Http::Policies::TransportPolicy>(transportAdapter);
+  std::make_unique<Azure::Core::Http::Policies::_internal::TransportPolicy>(transportAdapter);
 
     std::vector<std::unique_ptr<Azure::Core::Http::Policies::HttpPolicy>> policies;
     policies.emplace_back(std::move(transportPolicy));
@@ -152,7 +154,7 @@ namespace Azure { namespace Core { namespace Test {
 
     auto transportAdapter = std::make_shared<Azure::Core::Http::CurlTransport>(curlOptions);
     auto transportPolicy =
-  std::make_unique<Azure::Core::Http::Policies::TransportPolicy>(transportAdapter);
+  std::make_unique<Azure::Core::Http::Policies::_internal::TransportPolicy>(transportAdapter);
 
     std::vector<std::unique_ptr<Azure::Core::Http::Policies::HttpPolicy>> policies;
     policies.emplace_back(std::move(transportPolicy));
@@ -182,7 +184,8 @@ namespace Azure { namespace Core { namespace Test {
     auto transportAdapter = std::make_shared<Azure::Core::Http::CurlTransport>(curlOptions);
     Azure::Core::Http::Policies::TransportOptions options;
     options.Transport = transportAdapter;
-    auto transportPolicy = std::make_unique<Azure::Core::Http::Policies::TransportPolicy>(options);
+    auto transportPolicy
+        = std::make_unique<Azure::Core::Http::Policies::_internal::TransportPolicy>(options);
 
     std::vector<std::unique_ptr<Azure::Core::Http::Policies::HttpPolicy>> policies;
     policies.emplace_back(std::move(transportPolicy));
@@ -213,7 +216,8 @@ namespace Azure { namespace Core { namespace Test {
     auto transportAdapter = std::make_shared<Azure::Core::Http::CurlTransport>();
     Azure::Core::Http::Policies::TransportOptions options;
     options.Transport = transportAdapter;
-    auto transportPolicy = std::make_unique<Azure::Core::Http::Policies::TransportPolicy>(options);
+    auto transportPolicy
+        = std::make_unique<Azure::Core::Http::Policies::_internal::TransportPolicy>(options);
 
     std::vector<std::unique_ptr<Azure::Core::Http::Policies::HttpPolicy>> policies;
     policies.emplace_back(std::move(transportPolicy));
@@ -247,7 +251,8 @@ namespace Azure { namespace Core { namespace Test {
     auto transportAdapter = std::make_shared<Azure::Core::Http::CurlTransport>(curlOptions);
     Azure::Core::Http::Policies::TransportOptions options;
     options.Transport = transportAdapter;
-    auto transportPolicy = std::make_unique<Azure::Core::Http::Policies::TransportPolicy>(options);
+    auto transportPolicy
+        = std::make_unique<Azure::Core::Http::Policies::_internal::TransportPolicy>(options);
 
     {
       // use inner scope to remove the pipeline and make sure we don't keep the connection in the