+++ sdk/keyvault/azure-security-keyvault-common/test/ut/pipeline_test.cpp
@@ -15,7 +15,8 @@ using namespace Azure::Security::KeyVault::Common::_internal;
 TEST(KeyVaultPipeline, initPipeline)
 {
   std::vector<std::unique_ptr<Azure::Core::Http::Policies::HttpPolicy>> policies;
-  policies.emplace_back(std::make_unique<Azure::Core::Http::Policies::TransportPolicy>());
+  policies.emplace_back(
+      std::make_unique<Azure::Core::Http::Policies::_internal::TransportPolicy>());
   Azure::Core::Url url("urlTest");
   Azure::Core::_internal::ClientOptions options;
   Azure::Core::Http::_internal::HttpPipeline pipeline(