+++ sdk/storage/azure-storage-files-datalake/src/datalake_file_system_client.cpp
@@ -65,7 +65,7 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
         std::make_unique<_internal::StorageServiceVersionPolicy>(newOptions.ApiVersion));
     m_pipeline = std::make_shared<Azure::Core::Http::_internal::HttpPipeline>(
         newOptions,
-        _internal::FileServicePackageName,
+        _internal::DatalakeServicePackageName,
         PackageVersion::VersionString(),
         std::move(perRetryPolicies),
         std::move(perOperationPolicies));
@@ -89,14 +89,14 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
       Azure::Core::Credentials::TokenRequestContext tokenContext;
       tokenContext.Scopes.emplace_back(_internal::StorageScope);
       perRetryPolicies.emplace_back(
-          std::make_unique<Azure::Core::Http::Policies::BearerTokenAuthenticationPolicy>(
+          std::make_unique<Azure::Core::Http::Policies::_internal::BearerTokenAuthenticationPolicy>(
               credential, tokenContext));
     }
     perOperationPolicies.emplace_back(
         std::make_unique<_internal::StorageServiceVersionPolicy>(options.ApiVersion));
     m_pipeline = std::make_shared<Azure::Core::Http::_internal::HttpPipeline>(
         options,
-        _internal::FileServicePackageName,
+        _internal::DatalakeServicePackageName,
         PackageVersion::VersionString(),
         std::move(perRetryPolicies),
         std::move(perOperationPolicies));
@@ -118,7 +118,7 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
         std::make_unique<_internal::StorageServiceVersionPolicy>(options.ApiVersion));
     m_pipeline = std::make_shared<Azure::Core::Http::_internal::HttpPipeline>(
         options,
-        _internal::FileServicePackageName,
+        _internal::DatalakeServicePackageName,
         PackageVersion::VersionString(),
         std::move(perRetryPolicies),
         std::move(perOperationPolicies));
@@ -347,7 +347,46 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
       const RenameFileOptions& options,
       const Azure::Core::Context& context) const
   {
-    return this->GetDirectoryClient("").RenameFile(fileName, destinationFilePath, options, context);
+    std::string destinationFileSystem;
+    if (options.DestinationFileSystem.HasValue())
+    {
+      destinationFileSystem = options.DestinationFileSystem.GetValue();
+    }
+    else
+    {
+      const std::string& currentPath = m_fileSystemUrl.GetPath();
+      destinationFileSystem = currentPath.substr(0, currentPath.find('/'));
+    }
+
+    auto sourceDfsUrl = m_fileSystemUrl;
+    sourceDfsUrl.AppendPath(_internal::UrlEncodePath(fileName));
+
+    auto destinationDfsUrl = m_fileSystemUrl;
+    destinationDfsUrl.SetPath(_internal::UrlEncodePath(destinationFileSystem));
+    destinationDfsUrl.AppendPath(_internal::UrlEncodePath(destinationFilePath));
+
+    _detail::DataLakeRestClient::Path::CreateOptions protocolLayerOptions;
+    protocolLayerOptions.Mode = _detail::PathRenameMode::Legacy;
+    protocolLayerOptions.SourceLeaseId = options.SourceAccessConditions.LeaseId;
+    protocolLayerOptions.LeaseIdOptional = options.AccessConditions.LeaseId;
+    protocolLayerOptions.IfMatch = options.AccessConditions.IfMatch;
+    protocolLayerOptions.IfNoneMatch = options.AccessConditions.IfNoneMatch;
+    protocolLayerOptions.IfModifiedSince = options.AccessConditions.IfModifiedSince;
+    protocolLayerOptions.IfUnmodifiedSince = options.AccessConditions.IfUnmodifiedSince;
+    protocolLayerOptions.SourceIfMatch = options.SourceAccessConditions.IfMatch;
+    protocolLayerOptions.SourceIfNoneMatch = options.SourceAccessConditions.IfNoneMatch;
+    protocolLayerOptions.SourceIfModifiedSince = options.SourceAccessConditions.IfModifiedSince;
+    protocolLayerOptions.SourceIfUnmodifiedSince = options.SourceAccessConditions.IfUnmodifiedSince;
+    protocolLayerOptions.RenameSource = "/" + sourceDfsUrl.GetPath();
+    auto result = _detail::DataLakeRestClient::Path::Create(
+        destinationDfsUrl, *m_pipeline, context, protocolLayerOptions);
+
+    auto renamedBlobClient
+        = Blobs::BlobClient(_detail::GetBlobUrlFromUrl(destinationDfsUrl), m_pipeline);
+    auto renamedFileClient = DataLakeFileClient(
+        std::move(destinationDfsUrl), std::move(renamedBlobClient), m_pipeline);
+    return Azure::Response<DataLakeFileClient>(
+        std::move(renamedFileClient), result.ExtractRawResponse());
   }
 
   Azure::Response<DataLakeDirectoryClient> DataLakeFileSystemClient::RenameDirectory(
@@ -356,8 +395,46 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
       const RenameDirectoryOptions& options,
       const Azure::Core::Context& context) const
   {
-    return this->GetDirectoryClient("").RenameSubdirectory(
-        directoryName, destinationDirectoryPath, options, context);
+    std::string destinationFileSystem;
+    if (options.DestinationFileSystem.HasValue())
+    {
+      destinationFileSystem = options.DestinationFileSystem.GetValue();
+    }
+    else
+    {
+      const std::string& currentPath = m_fileSystemUrl.GetPath();
+      destinationFileSystem = currentPath.substr(0, currentPath.find('/'));
+    }
+
+    auto sourceDfsUrl = m_fileSystemUrl;
+    sourceDfsUrl.AppendPath(_internal::UrlEncodePath(directoryName));
+
+    auto destinationDfsUrl = m_fileSystemUrl;
+    destinationDfsUrl.SetPath(_internal::UrlEncodePath(destinationFileSystem));
+    destinationDfsUrl.AppendPath(_internal::UrlEncodePath(destinationDirectoryPath));
+
+    _detail::DataLakeRestClient::Path::CreateOptions protocolLayerOptions;
+    protocolLayerOptions.Mode = _detail::PathRenameMode::Legacy;
+    protocolLayerOptions.SourceLeaseId = options.SourceAccessConditions.LeaseId;
+    protocolLayerOptions.LeaseIdOptional = options.AccessConditions.LeaseId;
+    protocolLayerOptions.IfMatch = options.AccessConditions.IfMatch;
+    protocolLayerOptions.IfNoneMatch = options.AccessConditions.IfNoneMatch;
+    protocolLayerOptions.IfModifiedSince = options.AccessConditions.IfModifiedSince;
+    protocolLayerOptions.IfUnmodifiedSince = options.AccessConditions.IfUnmodifiedSince;
+    protocolLayerOptions.SourceIfMatch = options.SourceAccessConditions.IfMatch;
+    protocolLayerOptions.SourceIfNoneMatch = options.SourceAccessConditions.IfNoneMatch;
+    protocolLayerOptions.SourceIfModifiedSince = options.SourceAccessConditions.IfModifiedSince;
+    protocolLayerOptions.SourceIfUnmodifiedSince = options.SourceAccessConditions.IfUnmodifiedSince;
+    protocolLayerOptions.RenameSource = "/" + sourceDfsUrl.GetPath();
+    auto result = _detail::DataLakeRestClient::Path::Create(
+        destinationDfsUrl, *m_pipeline, context, protocolLayerOptions);
+
+    auto renamedBlobClient
+        = Blobs::BlobClient(_detail::GetBlobUrlFromUrl(destinationDfsUrl), m_pipeline);
+    auto renamedDirectoryClient = DataLakeDirectoryClient(
+        std::move(destinationDfsUrl), std::move(renamedBlobClient), m_pipeline);
+    return Azure::Response<DataLakeDirectoryClient>(
+        std::move(renamedDirectoryClient), result.ExtractRawResponse());
   }
 
 }}}} // namespace Azure::Storage::Files::DataLake