+++ sdk/storage/azure-storage-files-datalake/CHANGELOG.md
@@ -1,6 +1,12 @@
 # Release History
 
-## 12.0.0-beta.9 (Unreleased)
+## 12.0.0-beta.10 (Unreleased)
+
+### Bug Fixes
+
+- Rename functions always fail because `/` was left out in rename source path.
+
+## 12.0.0-beta.9 (2021-03-23)
 
 ### New Features
 
@@ -22,8 +28,9 @@
 - Removed `DataLake` from the names of return types and option types.
 - Removed `RequestId` from the return types.
 - Changed `BodyStream` parameter of `Append` function from pointer to reference.
-- Removed `PathRenameMode`, `PathGetPropertiesAction`, `PathSetAccessControlRecursiveMode`, `FileSystemResourceType` and `FileSystemResourceType`.
+- Removed `PathRenameMode`, `PathGetPropertiesAction`, `PathSetAccessControlRecursiveMode`, `FileSystemResourceType`, `PathExpiryOptions` and `FileSystemResourceType`.
 - Removed `IsAccessTierInferred` and `AccessTierChangedOn` from `PathProperties`.
+- Renamed `LeaseDurationType` to `LeaseDuration`, `LeaseStateType` to `LeaseState` and `LeaseStatusType` to `LeaseStatus`.
 
 ## 12.0.0-beta.8 (2021-02-12)
 
@@ -97,7 +104,7 @@
 - Renamed `DataLakePathClient::GetAccessControls` to `DataLakePathClient::GetAccessControlList`.
 - Removed unused parameters, options, results and functions in protocol layer.
 
-## 12.0.0-beta.6 (2020-01-14)
+## 12.0.0-beta.6 (2021-01-14)
 
 ### New Features